create connection session in SqlExecutionHandler
