+++ server/odc-service/src/main/java/com/oceanbase/odc/service/quartz/executor/AbstractQuartzJob.java
@@ -26,8 +26,8 @@ import com.oceanbase.odc.service.schedule.job.DataArchiveDeleteJob;
 import com.oceanbase.odc.service.schedule.job.DataArchiveJob;
 import com.oceanbase.odc.service.schedule.job.DataArchiveRollbackJob;
 import com.oceanbase.odc.service.schedule.job.DataDeleteJob;
-import com.oceanbase.odc.service.schedule.job.LogicalDatabaseChangeJob;
 import com.oceanbase.odc.service.schedule.job.LoadDataJob;
+import com.oceanbase.odc.service.schedule.job.LogicalDatabaseChangeJob;
 import com.oceanbase.odc.service.schedule.job.OdcJob;
 import com.oceanbase.odc.service.schedule.job.OnlineSchemaChangeCompleteJob;
 import com.oceanbase.odc.service.schedule.job.PartitionPlanJob;
